% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis2fun.R
\name{basis2fun}
\alias{basis2fun}
\alias{basis2fun,bspline_series,numeric-method}
\alias{basis2fun,Fourier_series,numeric-method}
\title{From the summation series of a functional basis to function value}
\usage{
basis2fun(object, x)

\S4method{basis2fun}{bspline_series,numeric}(object, x)

\S4method{basis2fun}{Fourier_series,numeric}(object, x)
}
\arguments{
\item{object}{An object that represents a functional basis.}

\item{x}{point(s) to take value.}
}
\value{
A numeric atomic vector.
See \code{\link{bsplineSeries2fun}} and \code{\link{FourierSeries2fun}}.
}
\description{
Generic function to compute function value from summation series of a functional basis.
}
\details{
When applied to \code{\link{bspline_series}} object, equivalent to \code{\link{bsplineSeries2fun}}.\cr
When applied to \code{\link{Fourier_series}} object, equivalent to \code{\link{FourierSeries2fun}}.
}
\author{
Heyang Ji
}
