% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.functional_variable.R
\name{dim,functional_variable-method}
\alias{dim,functional_variable-method}
\title{Extract dimensionality of functional data.}
\usage{
\S4method{dim}{functional_variable}(x)
}
\arguments{
\item{x}{a \code{\link{functional_variable}} object.}
}
\value{
Retruns a 2-element numeric vector.
}
\description{
Extract the dimensionality of slot X of functional_variable object.
}
\examples{
fv = functional_variable(X=array(rnorm(12),dim = 4:3),period = 3)
dim(fv)
}
\author{
Heyang Ji
}
