% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.beta.R
\name{fc.beta}
\alias{fc.beta}
\alias{fc.beta,fcRegression-method}
\alias{fc.beta,fcQR-method}
\title{Extract the value of coefficient parameter function}
\usage{
fc.beta(object, ...)

\S4method{fc.beta}{fcRegression}(object, FC = 1, t_points = NULL)

\S4method{fc.beta}{fcQR}(object, FC = 1, t_points = NULL)
}
\arguments{
\item{object}{An object that represents a functional covariates linear model.}

\item{...}{More arguments.}

\item{FC}{An integer, represent the ordinal number of the functional covariate.
Default is 1, which is take the first functional covariate.}

\item{t_points}{Sequence of the measurement (time) points.}
}
\value{
A numeric atomic vector
}
\description{
Generic function to extract the value of coefficient parameter function of the covariates
from linear model with functional covariates at some certain points.
}
\author{
Heyang Ji
}
