% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourierSeries2fun.R
\name{FourierSeries2fun}
\alias{FourierSeries2fun}
\alias{FourierSeries2fun,Fourier_series,numeric-method}
\title{Compute the value of the Fourier summation series at certain points.}
\usage{
FourierSeries2fun(object, t)

\S4method{FourierSeries2fun}{Fourier_series,numeric}(object, t)
}
\arguments{
\item{object}{an object of \code{\link{Fourier_series}} class.}

\item{t}{Value of t.}
}
\value{
A numeric atomic vector
}
\description{
Compute the value of the Fourier summation series at certain points.
}
\examples{
fsc = Fourier_series(
           double_constant = 0.5,
           cos = c(0,0.3),
           sin = c(1,0.7),
           k_cos = 1:2,
           )
          FourierSeries2fun(fsc,1:5)

}
\author{
Heyang Ji
}
