% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bspline_series.R
\name{plot,bspline_series-method}
\alias{plot,bspline_series-method}
\title{Plot b-splines baisi summation series.}
\usage{
\S4method{plot}{bspline_series}(x)
}
\arguments{
\item{x}{A \code{\link{bspline_series}} object.}
}
\value{
No return value. Generate a scatter plot.
}
\description{
Plot b-splines baisi summation series.
}
\examples{
bsb = bspline_basis(
Boundary.knots = c(0,24),
intercept      = TRUE,
df             = NULL,
degree         = 3
)
bss = bspline_series(
coef = c(2,1,1.5,3),
bspline_basis = bsb
)
plot(bss)
}
\author{
Heyang Ji
}
