% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Fourier_series.R
\name{plot,Fourier_series-method}
\alias{plot,Fourier_series-method}
\title{Plot Fourier basis summation series.}
\usage{
\S4method{plot}{Fourier_series}(x)
}
\arguments{
\item{x}{A \code{\link{Fourier_series}} object.}
}
\value{
No return value. Generate a scatter plot.
}
\description{
Plot Fourier basis summation series.
}
\examples{
fsc = Fourier_series(
double_constant = 0.5,
cos = c(0,0.3),
sin = c(1,0.7),
k_cos = 1:2,
)
plot(fsc)
}
\author{
Heyang Ji
}
