% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.numeric_series.R
\name{plot,numericBasis_series-method}
\alias{plot,numericBasis_series-method}
\title{Plot numeric basis function summation series.}
\usage{
\S4method{plot}{numericBasis_series}(x)
}
\arguments{
\item{x}{A \code{\link{numericBasis_series}} object.}
}
\description{
Plot numeric basis function summation series.
}
\examples{
t_0 = 0
period = 1
t_points = seq(0.05,0.95,length.out = 19)
nb = numeric_basis(
  basis_function = cbind(1/2,cos(2*pi*t_points),sin(2*pi*t_points)),
  t_points       = t_points,
  t_0            = t_0,
  period         = period
)
ns = numericBasis_series(coef = c(0.8,1.2,1.6),numeric_basis = nb)
plot(ns)
}
\author{
Heyang Ji
}
