% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fcQR.R
\name{predict.fcQR}
\alias{predict.fcQR}
\title{Predicted values based on fcQR object}
\usage{
\method{predict}{fcQR}(object, newData.FC, newData.Z = NULL, ...)
}
\arguments{
\item{object}{A fcQR class object produced by \code{\link{fcQR}}.}

\item{newData.FC}{A atomic vector or a matrix or a dataframe or a functional_variable class object or a list of objects above.
See argument FC in \code{\link{fcRegression}}.}

\item{newData.Z}{A dataframe or a matrix or a atomic vector. See argument Z in \code{\link{fcRegression}}.}

\item{...}{Further arguments passed to or from other methods \code{\link[quantreg]{predict.rq}}.}
}
\value{
See \code{\link[quantreg]{predict.rq}}.
}
\description{
Predicted values based on the Quantile linear model with functional covariates represented by a "fcQR" class object.
}
\details{
If no new data is input, will return the fitted value.
}
\author{
Heyang Ji
}
